requirements.txt
Django==5.0
psycopg2-binary==2.9.9
Pillow==10.2.0
django-crispy-forms==2.1
crispy-tailwind==1.0.3
qrcode[pil]==7.4.2
reportlab==4.1.0
django-environ==0.11.2
whitenoise==6.6.0
gunicorn==21.2.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
twilio==8.11.1
sendgrid==6.11.0
django-import-export==3.3.7
django-filter==23.5
djangorestframework==3.14.0
WeasyPrint==60.2
